expand_heredoc_document() を heredoc_expander.c に移動
