+++ Makefile
@@ -30,7 +30,7 @@ SRCS = cmd_cmd_invocation.c cmd_cmd_invocation2.c cmd_exec_command.c		\
 	lexer1.c lexer2.c lexer3.c lexer4.c parse1.c parse2.c parse3.c			\
 	parse_utils.c parse_utils2.c											\
 																			\
-	path.c g_cwd.c string_node2string.c expand_env_var.c					\
+	path.c g_cwd.c string_node2string.c env_expander.c heredoc_expander.c	\
 	expand_string_node.c split_expanded_str.c t_var.c t_var2.c				\
 																			\
 	builtin.c builtin_echo.c builtin_env.c builtin_exit.c builtin_cd.c		\
